+++ lib/serverkit/actions/diff.rb
@@ -3,13 +3,13 @@ require "thread"
 
 module Serverkit
   module Actions
-    class Check < Base
+    class Diff < Base
       def run
         backends.map do |backend|
           Thread.new do
             recipe.resources.map(&:dup).each do |resource|
               resource.backend = backend
-              result = resource.check ? "OK" : "NG"
+              result = resource.diff ? "OK" : "NG"
               puts "[ #{result} ] #{resource.id} on #{host_for(backend)}"
             end
           end